requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
sqlalchemy==2.0.27
pydantic==1.10.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.1.7
websockets==12.0
python-multipart==0.0.9
alembic==1.13.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
email-validator==2.1.0 